+++ test/bulk/bulk_test.exs
@@ -110,5 +110,120 @@ defmodule Snap.BulkTest do
                actions
                |> Bulk.perform(Cluster, @test_index, page_size: 2, page_wait: 10, max_errors: 0)
     end
+
+    test "upserting with doc_as_upsert" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      actions = [
+        %Action.Update{id: 1, doc: %{foo: "bar"}, doc_as_upsert: true}
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      assert {:ok, %{"_source" => %{"foo" => "bar"}}} =
+               Snap.Document.get(Cluster, @test_index, 1)
+    end
+
+    test "upserting with an explicit upsert document" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      actions = [
+        %Action.Update{id: 1, doc: %{foo: "bar"}, upsert: %{foo: "baz", count: 1}}
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      assert {:ok, %{"_source" => %{"foo" => "baz", "count" => 1}}} =
+               Snap.Document.get(Cluster, @test_index, 1)
+    end
+
+    test "updating with a script and no doc" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      actions = [
+        %Action.Index{id: 1, doc: %{count: 1}},
+        %Action.Update{
+          id: 1,
+          script: %{source: "ctx._source.count += params.by", params: %{by: 4}},
+          retry_on_conflict: 3
+        }
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      assert {:ok, %{"_source" => %{"count" => 5}}} = Snap.Document.get(Cluster, @test_index, 1)
+    end
+
+    test "updating with a scripted upsert" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      actions = [
+        %Action.Update{
+          id: 1,
+          script: %{source: "ctx._source.count = params.count", params: %{count: 7}},
+          upsert: %{},
+          scripted_upsert: true
+        }
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      assert {:ok, %{"_source" => %{"count" => 7}}} = Snap.Document.get(Cluster, @test_index, 1)
+    end
+
+    test "indexing with an external version" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      actions = [
+        %Action.Index{id: 1, doc: %{foo: "bar"}, version: 5, version_type: :external}
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      assert {:ok, %{"_version" => 5}} = Snap.Document.get(Cluster, @test_index, 1)
+
+      stale = [
+        %Action.Index{id: 1, doc: %{foo: "baz"}, version: 4, version_type: :external}
+      ]
+
+      assert {:error, %Snap.BulkError{errors: [error]}} =
+               Bulk.perform(stale, Cluster, @test_index)
+
+      assert error.type == "version_conflict_engine_exception"
+    end
+
+    test "indexing with if_seq_no and if_primary_term" do
+      {:ok, _} = Snap.Indexes.create(Cluster, @test_index, %{})
+
+      assert :ok == Bulk.perform([%Action.Index{id: 1, doc: %{foo: "bar"}}], Cluster, @test_index)
+
+      {:ok, %{"_seq_no" => seq_no, "_primary_term" => primary_term}} =
+        Snap.Document.get(Cluster, @test_index, 1)
+
+      actions = [
+        %Action.Index{
+          id: 1,
+          doc: %{foo: "baz"},
+          if_seq_no: seq_no,
+          if_primary_term: primary_term
+        }
+      ]
+
+      assert :ok == Bulk.perform(actions, Cluster, @test_index)
+
+      stale = [
+        %Action.Index{
+          id: 1,
+          doc: %{foo: "qux"},
+          if_seq_no: seq_no,
+          if_primary_term: primary_term
+        }
+      ]
+
+      assert {:error, %Snap.BulkError{errors: [error]}} =
+               Bulk.perform(stale, Cluster, @test_index)
+
+      assert error.type == "version_conflict_engine_exception"
+    end
   end
 end